Simple USART Example
This simple example illustrates the initialization of USART2 on
STM32F407 processors for a simple communications protocol. When it
receives 't' the USART transmits 'T' in response.

This project compiles with the Atollic TrueStudio IDE, but should with
little effort compile with most STM32 development platforms,, and runs
as is on the STM32F4-Discovery board